Backtesting program on Historical Data
Backtesting is a type of program that allows traders to test potential trading strategies using historical data. The program recreates the behaviour of trades and their reaction to a Zerodha trading strategy, and the resulting data can then be used to measure and optimise the effectiveness of a given strategy before applying it to real market conditions. Backtesting strategies work on the assumption that trades that have performed successfully in the past will perform well in the future